Update README.md
Hi John, I'm Dan from Deepnote. We're a small startup creating a better data science notebook and one of the things we're really interested is making it easier for people to try and learn from existing projects.
I'm quite new to this, so one of the first things I wanted to do was to try Deepnote with a couple of public repos, including yours. :)
Since I already set it up for myself, I thought it might be useful for others too to have a one-click button to run your repo, so I submitted a PR.

Thanks for your work, Dan